pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oracle-pg-sync-audit"
version = "0.1.0"
description = "Oracle to PostgreSQL sync, metadata audit, and DBA reporting toolkit."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "oracledb>=2.4.1",
  "psycopg[binary]>=3.2.3",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.2",
  "pandas>=2.2.3",
  "openpyxl>=3.1.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
]

[project.scripts]
oracle-pg-sync-audit = "oracle_pg_sync.cli:main"

[tool.setuptools.packages.find]
include = ["oracle_pg_sync*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
